Fix FFI

+++ peardrop_dart/lib/src/util.dart
@@ -11,16 +11,16 @@ void memcpy<T>(List<T> src, List<T> dst) {
   }
 }
 
-/// Copies a List<Uint8> into a C-heap buffer.
-Pointer<Uint8> cbcopy(List<Uint8> buffer) {
+/// Copies a List<int> into a C-heap buffer.
+Pointer<Uint8> cbcopy(List<int> buffer) {
   Pointer<Uint8> cbuffer = allocate(count: buffer.length);
   memcpy(buffer, cbuffer.asTypedList(buffer.length));
   return cbuffer;
 }
 
-/// Copies a C-heap buffer into a List<Uint8>.
-List<Uint8> bbcopy(Pointer<Uint8> ptr, int length) {
-  List<Uint8> buffer = List(length);
+/// Copies a C-heap buffer into a List<int>.
+List<int> bbcopy(Pointer<Uint8> ptr, int length) {
+  List<int> buffer = List(length);
   memcpy(ptr.asTypedList(length), buffer);
   return buffer;
 }
